requirements.txt
requests
bs4
cairosvg
PyPDF2
fpdf
Pillow
pyopenssl
progressbar
lxml
progressbar2
selenium